requirements.txt
boto3==1.35.10
botocore==1.35.10
jmespath==1.0.1
prometheus_client==0.20.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
